@@ -8,18 +8,12 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
 # Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
 # In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-# In Windows, build-name is used as the major, minor, and patch parts
-# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.9.2
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -54,6 +48,7 @@ dependencies:
   intl: ^0.20.2
   easy_localization: ^3.0.8
   flutter_native_splash: ^2.4.6
+  sign_in_button: ^4.0.1
   #An SVG rendering and widget library for Flutter, which allows painting
   flutter_svg: ^2.2.1
   #for adapting screen and font size
@@ -100,18 +95,27 @@ flutter:
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/to/asset-from-package
 
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # Fonts
+  # Custom fonts configuration
   fonts:
     - family: Cairo
       fonts:
-        - asset: assets/fonts/Cairo/Cairo-VariableFont_slnt,wght.ttf
-          # - asset: assets/fonts/Cairo-Bold.ttf
+        - asset: assets/fonts/Cairo/static/Cairo-Regular.ttf
+        - asset: assets/fonts/Cairo/static/Cairo-Bold.ttf
+          weight: 700
+        - asset: assets/fonts/Cairo/static/Cairo-SemiBold.ttf
           weight: 600
+        - asset: assets/fonts/Cairo/static/Cairo-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Cairo/static/Cairo-Light.ttf
+          weight: 300
+    - family: Story Script
+      fonts:
+        - asset: assets/fonts/Story_Script/StoryScript-Regular.ttf
+    - family: Vina Sans
+      fonts:
+        - asset: assets/fonts/Vina_Sans/VinaSans-Regular.ttf
+
+
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
