@@ -1,5 +1,26 @@
-class LoginResponse {
-  final String token;
-
-  LoginResponse({required this.token});
-}
+// import 'package:json_annotation/json_annotation.dart';
+// part 'login_response.g.dart';
+//
+// @JsonSerializable()
+// class LoginResponse {
+//   String? message;
+//   @JsonKey(name: 'data')
+//   UserData? userData;
+//   bool? status;
+//   int? code;
+//
+//   LoginResponse({this.message, this.userData, this.status, this.code});
+//
+//   factory LoginResponse.fromJson(Map<String, dynamic> json) => _$LoginResponseFromJson(json);
+// }
+//
+// @JsonSerializable()
+// class UserData {
+//   String? token;
+//   @JsonKey(name: 'username')
+//   String? userName;
+//
+//   UserData({this.token, this.userName});
+//
+//   factory UserData.fromJson(Map<String, dynamic> json) => _$UserDataFromJson(json);
+// }
