@@ -1,6 +1,12 @@
 class Routes {
-  static const String splashScreen = '/';
+  static const String splashScreen = '/splashScreen';
   static const String onBoardingScreen = '/onBoardingScreen';
   static const String loginScreen = '/loginScreen';
   static const String signUpScreen = '/signUpScreen';
+  static const String otpScreen = '/otpScreen';
+  static const String otpSuccessScreen = '/otpSuccessScreen';
+  static const String chatScreen = '/chatScreen';
+  static const String forgotPasswordScreen = '/forgotPassword';
+  static const String resetPasswordScreen = '/resetPassword';
+  static const String signUpSuccessScreen = '/signUpSuccess';
 }
