@@ -1,6 +1,19 @@
-class LoginRequestBody {
-  final String email;
-  final String password;
+// GENERATED CODE - DO NOT MODIFY BY HAND
 
-  LoginRequestBody({required this.email, required this.password});
-}
+part of 'login_request_body.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+LoginRequestBody _$LoginRequestBodyFromJson(Map<String, dynamic> json) =>
+    LoginRequestBody(
+      email: json['email'] as String,
+      password: json['password'] as String,
+    );
+
+Map<String, dynamic> _$LoginRequestBodyToJson(LoginRequestBody instance) =>
+    <String, dynamic>{
+      'email': instance.email,
+      'password': instance.password,
+    };
