@@ -1,5 +1,38 @@
-class LoginResponse {
-  final String token;
-
-  LoginResponse({required this.token});
-}
+//
+// import 'package:flutter_moblie_app/features/login/data/models/login_response.dart';
+// // GENERATED CODE - DO NOT MODIFY BY HAND
+//
+// // part of 'login_response.dart';
+//
+// // **************************************************************************
+// // JsonSerializableGenerator
+// // **************************************************************************
+//
+//
+// LoginResponse _$LoginResponseFromJson(Map<String, dynamic> json) =>
+//     LoginResponse(
+//       message: json['message'] as String?,
+//       userData: json['data'] == null
+//           ? null
+//           : UserData.fromJson(json['data'] as Map<String, dynamic>),
+//       status: json['status'] as bool?,
+//       code: json['code'] as int?,
+//     );
+//
+// Map<String, dynamic> _$LoginResponseToJson(LoginResponse instance) =>
+//     <String, dynamic>{
+//       'message': instance.message,
+//       'data': instance.userData,
+//       'status': instance.status,
+//       'code': instance.code,
+//     };
+//
+// UserData _$UserDataFromJson(Map<String, dynamic> json) => UserData(
+//   token: json['token'] as String?,
+//   userName: json['username'] as String?,
+// );
+//
+// Map<String, dynamic> _$UserDataToJson(UserData instance) => <String, dynamic>{
+//   'token': instance.token,
+//   'username': instance.userName,
+// };
