@@ -2,231 +2,294 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../core/helpers/spacing.dart';
+import '../../../core/routing/routes.dart';
 import '../../../core/theming/colors.dart';
 import '../../../core/theming/styles.dart';
 import '../../../core/widgets/app_text_button.dart';
 
-// ============================================
-// PURE UI SECTION - NO BUSINESS LOGIC HERE
-// ============================================
 class LoginScreen extends StatelessWidget {
   const LoginScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    // UI State Variables (for demo purposes)
     bool isLoading = false;
-    String? emailError;
     String? passwordError;
 
-    // Form Controllers (normally these would be in a controller/cubit)
     final emailController = TextEditingController();
     final passwordController = TextEditingController();
 
     return Scaffold(
-      backgroundColor: Colors.white,
-      body: Stack(
-        children: [
-          Container(
-            width: double.infinity,
-            height: double.infinity,
-            decoration: BoxDecoration(
-              gradient: RadialGradient(
-                center: Alignment(-0.7, -0.7), // Top-left quadrant
-                radius: 1.5,
-                colors: [
-                  ColorsManager.layerBlur1.withOpacity(0.4),
-                  ColorsManager.layerBlur1.withOpacity(0.1),
-                  Colors.transparent,
-                ],
-                stops: const [0.0, 0.3, 0.8],
+      body: SizedBox(
+        width: double.infinity,
+        height: double.infinity,
+        child: Stack(
+          children: [
+            // Full screen gradient overlay
+            Container(
+              width: double.infinity,
+              height: double.infinity,
+              decoration: BoxDecoration(
+                gradient: RadialGradient(
+                  center: Alignment(-0.7, -0.7), // Top-left quadrant
+                  radius: 1.5,
+                  colors: [
+                    ColorsManager.layerBlur1.withOpacity(0.4),
+                    ColorsManager.layerBlur1.withOpacity(0.1),
+                    Colors.transparent,
+                  ],
+                  stops: const [0.0, 0.3, 0.8],
+                ),
               ),
             ),
-          ),
-          // Bottom-right gradient overlay
-          Container(
-            width: double.infinity,
-            height: double.infinity,
-            decoration: BoxDecoration(
-              gradient: RadialGradient(
-                center: Alignment(0.7, 0.7), // Bottom-right quadrant
-                radius: 1.5,
-                colors: [
-                  ColorsManager.layerBlur2.withOpacity(0.4),
-                  ColorsManager.layerBlur2.withOpacity(0.1),
-                  Colors.transparent,
-                ],
-                stops: const [0.0, 0.3, 0.8],
+
+            // Bottom-right gradient overlay
+            Container(
+              width: double.infinity,
+              height: double.infinity,
+              decoration: BoxDecoration(
+                gradient: RadialGradient(
+                  center: Alignment(0.7, 0.7), // Bottom-right quadrant
+                  radius: 1.5,
+                  colors: [
+                    ColorsManager.layerBlur2.withOpacity(0.4),
+                    ColorsManager.layerBlur2.withOpacity(0.1),
+                    Colors.transparent,
+                  ],
+                  stops: const [0.1, 0.3, 0.8],
+                ),
               ),
             ),
-          ),
 
-          Container(
-            child: Padding(
-              padding: EdgeInsets.symmetric(horizontal: 30.w, vertical: 30.h),
-              child: SingleChildScrollView(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // Header Section
-                    Text(
-                      'Welcome Back',
-                      style: TextStyles.font24BlueBold,
+            Center(
+              child: Padding(
+                padding: EdgeInsets.all(24.0.w),
+                child: Container(
+                    width: double.infinity,
+                    padding: EdgeInsets.all(24.0.w),
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(16.0),
+                      boxShadow: [
+                        BoxShadow(
+                          color: Colors.black.withOpacity(0.1),
+                          blurRadius: 10,
+                          offset: const Offset(0, 4),
+                        ),
+                      ],
                     ),
-                    verticalSpace(8),
-                    Text(
-                      'We\'re excited to have you back, can\'t wait to see what you\'ve been up to since you last logged in.',
-                      style: TextStyles.font14GrayRegular,
-                    ),
-                    verticalSpace(36),
-
-                    // Login Form
-                    Form(
-                      // key: formKey, // Uncomment when implementing logic
+                    child: Directionality(
+                      textDirection: TextDirection.rtl,
                       child: Column(
+                        mainAxisSize: MainAxisSize.min,
+                        crossAxisAlignment: CrossAxisAlignment.center,
                         children: [
-                          // Email Field
-                          TextFormField(
-                            controller: emailController,
-                            decoration: InputDecoration(
-                              labelText: 'Email',
-                              errorText: emailError,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                            ),
-                            keyboardType: TextInputType.emailAddress,
+                          verticalSpace(10),
+                          Image.asset(
+                            'assets/images/splash-logo.png',
+                            width: 80.w,
+                            height: 80.h,
                           ),
-                          verticalSpace(16),
-
-                          // Password Field
-                          TextFormField(
-                            controller: passwordController,
-                            decoration: InputDecoration(
-                              labelText: 'Password',
-                              errorText: passwordError,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                              suffixIcon: IconButton(
-                                icon: const Icon(Icons.visibility_off),
-                                onPressed: () {
-                                  // Toggle password visibility
-                                },
-                              ),
-                            ),
-                            obscureText: true,
-                          ),
-
-                          verticalSpace(24),
-
-                          // Forgot Password
-                          Align(
-                            alignment: Alignment.centerRight,
-                            child: Text(
-                              'Forgot Password?',
-                              style: TextStyles.font13BlueRegular,
-                            ),
-                          ),
-
-                          verticalSpace(40),
-
-                          // Login Button
-                          SizedBox(
-                            width: double.infinity,
-                            child: AppTextButton(
-                              buttonText: "Login",
-                              textStyle: TextStyles.font16WhiteSemiBold,
-                              onPressed: () {},
-                            ),
-                          ),
-
-                          verticalSpace(16),
-
-                          // Terms & Conditions
+                          Text(' تسجيل دخول', style: TextStyles.font24BlueBold,),
+                          verticalSpace(8),
                           Text(
-                            'By logging in, you agree to our Terms & Conditions and Privacy Policy',
-                            style: TextStyles.font13GrayRegular,
-                            textAlign: TextAlign.center,
+                            'ادخل الايميل او رقم الهاتف و كلمه المرور',
+                            style: TextStyles.font14GrayRegular,
+                            textAlign: TextAlign.right,
                           ),
-
-                          verticalSpace(24),
-
-                          // Don't have an account?
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Text(
-                                'Don\'t have an account?',
-                                style: TextStyles.font13DarkBlueRegular,
-                              ),
-                              TextButton(
-                                onPressed: () {
-                                  // Navigate to signup
-                                },
-                                child: Text(
-                                  'Sign Up',
-                                  style: TextStyles.font13BlueSemiBold,
+                          verticalSpace(12),
+                          // Login Form
+                          Form(
+                            child: Column(
+                              children: [
+                                TextFormField(
+                                  controller: emailController,
+                                  decoration: InputDecoration(
+                                    labelText: 'البريد الإلكتروني',
+                                    // errorText: emailError,
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8),
+                                    ),
+                                  ),
+                                  keyboardType: TextInputType.emailAddress,
                                 ),
-                              ),
-                            ],
+                                verticalSpace(16),
+                                TextFormField(
+                                  controller: passwordController,
+                                  decoration: InputDecoration(
+                                    labelText: 'كلمة المرور',
+                                    errorText: passwordError,
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8),
+                                    ),
+                                    suffixIcon: IconButton(
+                                      icon: const Icon(Icons.visibility_off),
+                                      onPressed: () {
+                                        // Toggle password visibility
+                                      },
+                                    ),
+                                  ),
+                                  obscureText: true,
+                                ),
+                                verticalSpace(5),
+                                // Forgot Password
+                                Row(
+                                  mainAxisAlignment: MainAxisAlignment
+                                      .spaceBetween,
+                                  children: [
+                                    Text(
+                                      'هل نسيت كلمة المرور؟',
+                                      style: TextStyles.font13BlueRegular,
+                                    ),
+                                    Row(
+                                      children: [
+
+                                        Transform.scale(
+                                          scale: 0.9,
+                                          child: Checkbox(
+                                            value: false,
+                                            // You'll need to manage this state
+                                            onChanged: (bool? value) {
+                                              // Handle checkbox state change
+                                            },
+                                            activeColor: ColorsManager.mainBlue,
+                                          ),
+                                        ), Text(
+                                          'تذكرني',
+                                          style: TextStyles
+                                              .font13DarkBlueRegular,
+                                        ),
+                                      ],
+                                    ),
+                                  ],
+                                ),
+                                verticalSpace(10),
+                                // Login Button
+                                SizedBox(
+                                  width: double.infinity,
+                                  child: AppTextButton(
+                                      buttonText: "تسجيل الدخول",
+                                      onPressed: () {
+                                        Navigator.of(
+                                          context,
+                                        ).pushNamed(Routes.chatScreen);
+                                      }, textStyle: TextStyle(color: Colors.white)
+                                  ),
+                                ),
+                                verticalSpace(10),
+                                // Terms & Conditions
+                                Text(
+                                  'بالدخول، أنت توافق على الشروط والأحكام.',
+                                  style: TextStyles.font13GrayRegular,
+                                  textAlign: TextAlign.center,
+                                ),
+                                verticalSpace(6),
+                                Padding(
+                                  padding: const EdgeInsets.symmetric(vertical: 16.0),
+                                  child: Row(
+                                    children: [
+                                      Expanded(
+                                        child: Divider(
+                                          color: Colors.grey[400],
+                                          thickness: 1,
+                                        ),
+                                      ),
+                                      Padding(
+                                        padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                                        child: Text('أو', style: TextStyle(color: Colors.grey[600])),
+                                      ),
+                                      Expanded(
+                                        child: Divider(
+                                          color: Colors.grey[400],
+                                          thickness: 1,
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                                const SizedBox(height: 16),
+                                Row(
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: [
+                                    Container(
+                                      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                                      decoration: BoxDecoration(
+                                        border: Border.all(color: Colors.grey[300]!),
+                                        borderRadius: BorderRadius.circular(12),
+                                      ),
+                                      child: Text(
+                                        'Google',
+                                        style: TextStyle(
+                                          color: Colors.grey[800],
+                                          fontWeight: FontWeight.w500,
+                                        ),
+                                      ),
+                                    ),
+                                    const SizedBox(width: 16),
+                                    Container(
+                                      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+                                      decoration: BoxDecoration(
+                                        border: Border.all(color: Colors.grey[300]!),
+                                        borderRadius: BorderRadius.circular(12),
+                                      ),
+                                      child: Text(
+                                        'Facebook',
+                                        style: TextStyle(
+                                          color: Colors.grey[800],
+                                          fontWeight: FontWeight.w500,
+                                        ),
+                                      ),
+                                    ),
+                                    const SizedBox(width: 16),
+                                    Container(
+                                      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                                      decoration: BoxDecoration(
+                                        border: Border.all(color: Colors.grey[300]!),
+                                        borderRadius: BorderRadius.circular(12),
+                                      ),
+                                      child: Text(
+                                        'Apple',
+                                        style: TextStyle(
+                                          color: Colors.grey[800],
+                                          fontWeight: FontWeight.w500,
+                                        ),
+                                      ),
+                                    ),
+                                  ],
+                                ),
+
+
+                                Row(
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: [
+                                    Text(
+                                      'ليس لديك حساب؟',
+                                      style: TextStyles.font13DarkBlueRegular,
+                                    ),
+                                    TextButton(
+                                      onPressed: () {
+                                        Navigator.of(
+                                          context,
+                                        ).pushNamed(Routes.signUpScreen);
+                                      },
+                                      child: Text(
+                                        'إنشاء حساب',
+                                        style: TextStyles.font13BlueSemiBold,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                              ],
+                            ),
                           ),
                         ],
                       ),
                     ),
-                  ],
-                ),
+                  ),
               ),
             ),
-          ),
-        ],
+          ],
+        ),
       ),
     );
   }
 }
-
-// ============================================
-// LOGIC SECTION - TO BE IMPLEMENTED LATER
-// ============================================
-/*
-// 1. State Management (using Cubit)
-class LoginCubit extends Cubit<LoginState> {
-  final formKey = GlobalKey<FormState>();
-  final emailController = TextEditingController();
-  final passwordController = TextEditingController();
-
-  // Add login logic here
-}
-
-// 2. State Classes
-abstract class LoginState {}
-class LoginInitial extends LoginState {}
-class LoginLoading extends LoginState {}
-class LoginSuccess extends LoginState {}
-class LoginError extends LoginState {
-  final String message;
-  LoginError(this.message);
-}
-
-// 3. Form Validation
-bool validateForm(GlobalKey<FormState> formKey) {
-  // Add form validation logic
-  return formKey.currentState?.validate() ?? false;
-}
-
-// 4. API Calls
-Future<void> loginUser(String email, String password) async {
-  // Add API call logic
-}
-*/
-
-// ============================================
-// WIDGETS SECTION - SEPARATE FILES
-// ============================================
-/*
-// Move these to separate widget files:
-- EmailAndPassword()
-- TermsAndConditionsText()
-- DontHaveAccountText()
-- LoginBlocListener()
-*/
